+++ lib/features/player/logic/session_extensions.dart
@@ -32,7 +32,6 @@ extension PlaybackSessionX on PlaybackSession {
       jsonChapters = chapters
           .map(
             (c) => Chapter(
-              displayIndex: c.id,
               start: c.start,
               end: c.end,
               title: c.title,
@@ -45,7 +44,6 @@ extension PlaybackSessionX on PlaybackSession {
           audioTracks
               ?.map(
                 (t) => Chapter(
-                  displayIndex: t.index,
                   start: t.startOffset,
                   end: t.startOffset + t.duration,
                   title: displayTitle ?? l10n.noTitle,